Refactors ASM instruction translators into sub-package
Moves ASM instruction translator modules into a new `instruction_translators` sub-package to improve code organization within the assembly generation module.
Updates all relevant import paths to reflect the new structure.

Additionally, initializes the `test_asm_translators` directory as a Python package.
